changes on cycling.py

+++ cycling.py
@@ -6,13 +6,13 @@ class Cycling(PhysicalActivity):
         self.distance = distance  
 
     def calculate_calories_burned(self):
-        distance_km = self.distance * 1.60934
-        weight_kg = self.weight * 0.453592
+        distance_km = int(self.distance * 1.60934)
+        weight_kg = int(self.weight * 0.453592)
 
         average_speed_kmh = 15
         calories_burned_per_hour_per_kg = 8.5
         total_hours = distance_km / average_speed_kmh
 
-        total_calories_burned = calories_burned_per_hour_per_kg * weight_kg * total_hours
+        total_calories_burned = int(calories_burned_per_hour_per_kg )* weight_kg * total_hours
 
         return int(total_calories_burned)
